* Options moved to Kconfig files:
  CONFIG_POWER
  CONFIG_POWER_I2C
  CONFIG_POWER_PFUZE100
  CONFIG_SPL_MAX_SIZE
  CONFIG_SPL_PAD_TO
  CONFIG_SPL_STACK
  CONFIG_SYS_LOAD_ADDR
  CONFIG_SYS_NS16550_SERIAL
  CONFIG_SYS_PBSIZE
  CONFIG_MTDIDS_DEFAULT
  CONFIG_MTDPARTS_DEFAULT

* Renamed options and function:
  CONFIG_DW_WDT_CLOCK_KHZ to CFG_DW_WDT_CLOCK_KHZ
  CONFIG_EXTRA_ENV_SETTINGS to CFG_EXTRA_ENV_SETTINGS
  CONFIG_HW_WATCHDOG to CONFIG_WDT
  CONFIG_POWER_PFUZE100_I2C_ADDR to  CFG_POWER_PFUZE100_I2C_ADDR
  CONFIG_SYS_SDRAM_BASE to CFG_SYS_SDRAM_BASE
  CONFIG_SYS_TEXT_BASE to CONFIG_TEXT_BASE
  CONFIG_SYS_NS16550_CLK to CFG_POWER_PFUZE100_I2C_ADDR
  CONFIG_SYS_NS16550_COM* to CFG_SYS_NS16550_COM*
  phy_get_interface_by_name to dev_read_phy_mode

* Added options:
  CONFIG_CUSTOM_SYS_INIT_SP_ADDR
  CONFIG_DESIGNWARE_WATCHDOG - for salutepmukf_defconfig
  CONFIG_HAS_CUSTOM_SYS_INIT_SP_ADDR
  CONFIG_POWER_LEGACY
  CONFIG_SPL_GPIO
  CONFIG_SPL_HAVE_INIT_STACK
  CONFIG_SPL_I2C
  CONFIG_SPL_NO_BSS_LIMIT
  CONFIG_SPL_POWER_DOMAIN
  CONFIG_SPL_SERIAL
  CONFIG_SPL_SHARES_INIT_SP_ADDR is not set, forced disabling required
  CONFIG_SPL_SYS_I2C_LEGACY
  CONFIG_SYS_I2C_LEGACY

* Removed options:
  CONFIG_SPL_I2C_SUPPORT
  CONFIG_SPL_SERIAL_SUPPORT

* common.h wsa removed, so some of the header files that were in it were
  added directly

* BOOTENV_DEV_UBUFS was changed, becase error occured:
    macro "BOOTENV_DEV_UBIFS" requires 6 arguments, but only 3 given

Issue: #MCOM02SW-2400
Depends-On: https://gerrit.elvees.com/c/mcom02/tests/+/54333

+++ examples/api/glue.c
@@ -3,7 +3,6 @@
  * (C) Copyright 2007-2008 Semihalf, Rafal Jaworowski <raj@semihalf.com>
  */
 
-#include <common.h>
 #include <env.h>
 #include <linux/types.h>
 #include <api_public.h>
@@ -42,8 +41,8 @@ static int valid_sig(struct api_signature *sig)
 int api_search_sig(struct api_signature **sig)
 {
 	unsigned char *sp;
-	uint32_t search_start = 0;
-	uint32_t search_end = 0;
+	uintptr_t search_start = 0;
+	uintptr_t search_end = 0;
 
 	if (sig == NULL)
 		return 0;
@@ -153,7 +152,6 @@ unsigned long ub_get_timer(unsigned long base)
 
 	return cur;
 }
-
 
 /****************************************************************************
  *
@@ -365,11 +363,6 @@ const char * ub_env_enum(const char *last)
 
 	env = NULL;
 
-	/*
-	 * It's OK to pass only the name piece as last (and not the whole
-	 * 'name=val' string), since the API_ENUM_ENV call uses env_match()
-	 * internally, which handles such case
-	 */
 	if (!syscall(API_ENV_ENUM, NULL, last, &env))
 		return NULL;
